requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

asgiref==3.9.1
brotli==1.1.0
certifi==2025.8.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6; platform_system == "Windows"
dj-database-url==3.0.1
django==5.2.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
graphviz==0.21
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyjwt==2.10.1
python-dotenv==1.1.1
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == "win32, darwin"
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2; sys_platform == "win32"
urllib3==2.5.0
uvicorn==0.35.0
whitenoise[brotli]==6.9.0
